# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Accident:
    properties:
      aircraft_damage_description:
        type: string
      aircraft_id:
        type: integer
      aircraft_missing_flag:
        type: string
      entry_date:
        type: string
      event_local_date:
        type: string
      event_local_time:
        type: string
      event_type_description:
        type: string
      far_part:
        type: string
      fatal_flag:
        type: string
      flight_activity:
        type: string
      flight_number:
        type: string
      flight_phase:
        type: string
      fsdo_description:
        type: string
      id:
        type: integer
      location_id:
        type: integer
      remark_text:
        type: string
      updated:
        type: string
    type: object
  models.AccidentPaginatedResponse:
    properties:
      accidents:
        items:
          $ref: '#/definitions/models.Accident'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  models.Aircraft:
    properties:
      aircraft_make_name:
        type: string
      aircraft_model_name:
        type: string
      aircraft_operator:
        type: string
      id:
        type: integer
      registration_number:
        type: string
    type: object
  models.AircraftImage:
    properties:
      aircraft_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      s3_url:
        type: string
    type: object
  models.AircraftPaginatedResponse:
    properties:
      aircrafts:
        items:
          $ref: '#/definitions/models.Aircraft'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  models.ImagesForAircraftResponse:
    properties:
      aircraft_id:
        type: integer
      images:
        items:
          $ref: '#/definitions/models.AircraftImage'
        type: array
    type: object
  models.Injury:
    properties:
      accident_id:
        type: integer
      count:
        type: integer
      id:
        type: integer
      injury_severity:
        type: string
      person_type:
        type: string
    type: object
info:
  contact: {}
  description: API server for airaccidentdata.com
  title: AirAccidentData API
  version: "1.0"
paths:
  /accidents:
    get:
      description: Get a list of all aviation accidents
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of accidents per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccidentPaginatedResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get a list of accidents
      tags:
      - Accidents
  /accidents/{id}:
    get:
      description: Retrieve details of an accident by its ID
      parameters:
      - description: Accident ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aircraft'
        "400":
          description: Invalid accident ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Accident not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get an accident by ID
      tags:
      - Accidents
  /aircrafts:
    get:
      description: Retrieve a list of all aircrafts.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of aircraft per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Aircrafts data with pagination details
          schema:
            $ref: '#/definitions/models.AircraftPaginatedResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get a list of aircrafts
      tags:
      - Aircrafts
  /aircrafts/{id}:
    get:
      description: Retrieve details of an aircraft by its ID
      parameters:
      - description: Aircraft ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aircraft'
        "400":
          description: Invalid aircraft ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Aircraft not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get details about an aircraft by ID
      tags:
      - Aircrafts
  /aircrafts/{id}/images:
    get:
      description: Retrieve all images associated with a specific aircraft.
      parameters:
      - description: Aircraft ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Image IDs, Image URLs, and S3 URLs
          schema:
            $ref: '#/definitions/models.ImagesForAircraftResponse'
        "400":
          description: Invalid aircraft ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Aircraft not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all images for an aircraft
      tags:
      - Aircrafts
  /injuries/{id}:
    get:
      description: Retrieve injuries for an accident.
      parameters:
      - description: Aircraft ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Injury'
        "400":
          description: Invalid accident ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Accident not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get injuries for an accident
      tags:
      - Injuries
swagger: "2.0"